CV-456 Track.objectId values are negative (#84)
Track IDs in virtual camera are now int64, but in Connect SDK they are int.
Switched to using int64 as track ID in Connect SDK and sending track ID in JSON
as a string, rather than a number (integer number can only have 53 bits).

+++ include/domain-types.h
@@ -259,13 +259,13 @@ typedef std::vector<TrackPoint> TrackPoints;
 
 struct Track
 {
-    Track(int objectId, const timestamp_t& timestamp)
+    Track(int64_t objectId, const timestamp_t& timestamp)
         : objectId(objectId)
         , timestamp(timestamp)
     {
     }
 
-    int objectId;
+    int64_t objectId;
     timestamp_t timestamp;
     TrackPoints points;
 };
